requirements.txt
lxml==5.2.2
opcua==0.98.13
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
websockets==12.0
